requirements.txt
asgiref==3.9.2
attrs==25.3.0
black==24.10.0
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
flake8==7.1.1
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.3.0
platformdirs==4.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.23
pyflakes==3.2.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
-e git+https://github.com/mesuef1974/sh_school_015.git@6ea31d4492e06691ae961a842cd947253fefcf2c#egg=sh_school
six==1.17.0
sqlparse==0.5.3
tabula-py==2.10.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.37.0
Werkzeug==3.1.3
redis==5.2.1
django-rq==2.10.3
whitenoise==6.7.0